finale v. commented

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: girizzi <girizzi@student.42roma.it>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/03 18:55:31 by girizzi           #+#    #+#             */
-/*   Updated: 2025/02/06 19:57:09 by girizzi          ###   ########.fr       */
+/*   Updated: 2025/02/07 16:30:46 by girizzi          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,13 +18,14 @@
 # include <stdarg.h>
 # include <unistd.h>
 # include <stdint.h>
+# include <stdio.h>
 
 int	ft_printf(const char *format, ...);
 int	ft_print_char(char c);
 int	ft_print_str(char *s);
 int	ft_print_nbr(int n);
 int	ft_print_unsigned(unsigned int n);
-int	ft_print_hex(unsigned int num, int uppercase);
+int	ft_print_hex(unsigned long num, int uppercase);
 int	ft_print_ptr(void *ptr);
 
 #endif
